Fix Windows WaE

+++ src/megaclient.cpp
@@ -21682,7 +21682,7 @@ string KeyManager::tagHeader(const byte tag, size_t len) const
     vector<byte> res;
 
     res.push_back(tag);
-    res.push_back((len & 0xFF0000) >> 16);
+    res.push_back(static_cast<byte>((len & 0xFF0000) >> 16));
     res.push_back((len & 0xFF00) >> 8);
     res.push_back(len & 0xFF);
